requirements.txt
torch
torchvision
segmentation-models-pytorch
albumentations
imgaug
numpy
opencv-python
matplotlib
pandas
scikit-learn
PyYAML
